Update 7)random_number_generator.cpp
call rand() with the modulo operator to generate a random number between 0 and your favorite number. For example, rand() % 29 would output a random number between 0 and 28.

+++ 7)random_number_generator.cpp
@@ -6,6 +6,6 @@ int main() {
   srand (time(NULL));
   
   // Use rand() below to initialize the_amazing_random_number
-  double random_number=rand()%29;
+  int random_number=rand()%29;
   std::cout<<random_number;
 }
